# apps/vuln-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title:  TestBed API (api-demo)
  version: 1.0.0
  description: >-
    Static OpenAPI spec for the api-demo app. For full narrative docs see
    API-DOCUMENTATION.md. Endpoints are intentionally vulnerable for testing.
servers:
  - url: http://localhost:8880
...
paths:
...
  /api/v1/ecommerce/cart/add:
    post:
      tags: [Ecommerce]
      summary: "Cart add (price tampering)"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/cart/apply-discount:
    post:
      tags: [Ecommerce]
      summary: Discount stacking
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/cart/checkout:
    post:
      tags: [Ecommerce]
      summary: "Checkout (race/discount abuse)"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/cart/{cart_id}:
    get:
      tags: [Ecommerce]
      summary: "Cart details (IDOR)"
      parameters:
        - name: cart_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/chargebacks/submit:
    post:
      tags: [Ecommerce]
      summary: Chargeback fraud
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/checkout/complete:
    post:
      tags: [Ecommerce]
      summary: Payment bypass
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/checkout/submit:
    post:
      tags: [Ecommerce]
      summary: "Checkout submit (race)"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/v1/ecommerce/customers/{customer_id}/email:
    put:
      tags: [Ecommerce]
      summary: Account takeover
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/customers/{customer_id}/wishlist:
    get:
      tags: [Ecommerce]
      summary: Wishlist scraping
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/gift-cards/generate:
    post:
      tags: [Ecommerce]
      summary: Gift card abuse
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/gift-cards/{code}/balance:
    get:
      tags: [Ecommerce]
      summary: Gift card balance scraping
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/inventory/reserve:
    post:
      tags: [Ecommerce]
      summary: Reservation abuse
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/inventory/{item_id}:
    put:
      tags: [Ecommerce]
      summary: Inventory manipulation
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/loyalty/points/transfer:
    post:
      tags: [Ecommerce]
      summary: Points transfer abuse
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/loyalty/redeem:
    put:
      tags: [Ecommerce]
      summary: Redemption abuse
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/loyalty/tiers:
    put:
      tags: [Ecommerce]
      summary: Tier manipulation
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/v1/ecommerce/orders/{order_id}:
    get:
      tags: [Ecommerce]
      summary: "Order details (IDOR)"
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/orders/{order_id}/refund:
    post:
      tags: [Ecommerce]
      summary: Refund fraud
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/payment-methods/{method_id}:
    get:
      tags: [Ecommerce]
      summary: Payment method IDOR
      parameters:
        - name: method_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/ecommerce/pricing/override:
    put:
      tags: [Ecommerce]
      summary: "Pricing override (tampering)"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/v1/ecommerce/returns/approve:
    put:
      tags: [Ecommerce]
      summary: Approval bypass
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/returns/request:
    post:
      tags: [Ecommerce]
      summary: Return abuse
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/vendors/payouts:
    post:
      tags: [Ecommerce]
      summary: Payout manipulation
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/vendors/register:
    post:
      tags: [Ecommerce]
      summary: Vendor onboarding bypass
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/ecommerce/vendors/{vendor_id}:
    get:
      tags: [Ecommerce]
      summary: Vendor portal IDOR
      parameters:
        - name: vendor_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/cart/add:
    post:
      tags: [Ecommerce]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/cart/update:
    put:
      tags: [Ecommerce]
      summary: Put
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/inventory/check:
    post:
      tags: [Ecommerce]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/inventory/reserve:
    post:
      tags: [Ecommerce]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/products/listings:
    post:
      tags: [Ecommerce]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/ratings/bulk:
    post:
      tags: [Ecommerce]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/reviews/submit:
    post:
      tags: [Ecommerce]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/v1/ecommerce/vendor/webhooks/register:
    post:
      tags: [Ecommerce]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/vendors/auth/takeover:
    post:
      tags: [Ecommerce]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/vendors/backdoor:
    post:
      tags: [Ecommerce]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/vendors/documents/upload:
    post:
      tags: [Ecommerce]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/vendors/inventory/sabotage:
    post:
      tags: [Ecommerce]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/vendors/privileges/escalate:
    put:
      tags: [Ecommerce]
      summary: Put
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/vendors/register:
    post:
      tags: [Ecommerce]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...